[Refactor/#47]: mutation 커스텀 훅으로 변경
