+++ src/features/Memo/index.ts
@@ -1,3 +1,4 @@
+export { useAddMemo } from './hooks/useAddMemo';
 export { default as MemoDetail } from './components/MemoDetail';
 export { default as MemoForm } from './components/MemoForm/MemoForm';
 export { default as MemoList } from './components/MemoList/MemoList';
