+++ src/features/Review/components/ReviewForm/NewReviewForm.tsx
@@ -1,23 +1,20 @@
 import React from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useSplash } from '@/hooks/popup/useSplash';
-import { useStackMotionActions } from '@/store/stackMotionStore';
 import { splash } from '@/data/ui/splash';
 import Splash from '@/components/Splash/Splash';
-import { useMutation } from '@tanstack/react-query';
-import { useSession } from 'next-auth/react';
 import ReviewForm from './ReviewForm';
 import { ReviewForm as ReviewFormType } from '../../types/review';
-import { addNewReview } from '../../api/review.api';
+import { useAddReview } from '../../hooks/useAddReview';
 
 interface Props {
   isbn: string;
 }
 
 function NewReviewForm({ isbn }: Props) {
-  const { data: session } = useSession();
   const { isOpen, setIsOpen } = useSplash(splash.review.route);
-  const { setNewReviewId } = useStackMotionActions();
+  const { handleAddReview } = useAddReview(isbn, () => setIsOpen(true));
+
   const methods = useForm<ReviewFormType>({
     mode: 'onBlur',
     defaultValues: {
@@ -41,27 +38,6 @@ function NewReviewForm({ isbn }: Props) {
     !watch('pros').length ||
     !watch('cons').length ||
     !isValid;
-
-  const { mutate: handleAddReview } = useMutation({
-    mutationFn: (data: ReviewFormType) => {
-      const reqData = {
-        writer: session!.user.id,
-        isbn,
-        tags_pos: data.pros,
-        tags_neg: data.cons,
-        title: data.oneLiner,
-        content: data.review,
-        rating: data.rating!,
-      };
-
-      const result = addNewReview(reqData);
-      return result;
-    },
-    onSuccess: (result) => {
-      setNewReviewId(result.id!);
-      setIsOpen(true);
-    },
-  });
 
   return (
     <FormProvider {...methods}>
